+++ src/Budget.php
@@ -414,10 +414,11 @@ class Budget extends CommonDropdown
                     case 'Change':
                         $costtable = getTableForItemType($item->getType() . 'Cost');
 
-                        $sum = new QueryExpression(
-                            "SUM(" . $DB->quoteName("$costtable.actiontime") . " * " . $DB->quoteName("$costtable.cost_time") . "/" . HOUR_TIMESTAMP . "
-                                          + " . $DB->quoteName("$costtable.cost_fixed") . "
-                                          + " . $DB->quoteName("$costtable.cost_material") . ") AS " . $DB->quoteName('value')
+                        $sum = QueryFunction::sum(
+                            expression: $DB::quoteName("$costtable.actiontime") . " * " . $DB::quoteName("$costtable.cost_time") . "/" . HOUR_TIMESTAMP . "
+                                          + " . $DB::quoteName("$costtable.cost_fixed") . "
+                                          + " . $DB::quoteName("$costtable.cost_material"),
+                            alias: $DB::quoteName('value')
                         );
                         $criteria = [
                             'SELECT'       => [
@@ -756,11 +757,13 @@ class Budget extends CommonDropdown
                 case 'Problem':
                 case 'Change':
                     $costtable   = getTableForItemType($item->getType() . 'Cost');
-                    $sum = new QueryExpression(
-                        "SUM(" . $DB->quoteName("$costtable.actiontime") . " * " . $DB->quoteName("$costtable.cost_time") . "/" . HOUR_TIMESTAMP . "
-                                       + " . $DB->quoteName("$costtable.cost_fixed") . "
-                                       + " . $DB->quoteName("$costtable.cost_material") . ") AS " . $DB->quoteName('sumvalue')
+                    $sum = QueryFunction::sum(
+                        expression: $DB::quoteName("$costtable.actiontime") . " * " . $DB::quoteName("$costtable.cost_time") . "/" . HOUR_TIMESTAMP . "
+                                           + " . $DB::quoteName("$costtable.cost_fixed") . "
+                                           + " . $DB::quoteName("$costtable.cost_material"),
+                        alias: $DB::quoteName('sumvalue')
                     );
+
                     $criteria = [
                         'SELECT'       => [
                             $item->getTable() . '.entities_id',