+++ src/DeviceHardDrive.php
@@ -270,10 +270,10 @@ class DeviceHardDrive extends CommonDevice
             'width'              => 1000,
             'massiveaction'      => false,
             'joinparams'         => $main_joinparams,
-            'computation'        =>
-            '(SUM(' . $DB->quoteName('TABLE.capacity') . ') / COUNT(' .
-            $DB->quoteName('TABLE.id') . '))
-            * COUNT(DISTINCT ' . $DB->quoteName('TABLE.id') . ')',
+            'computation'        => '(' .
+                QueryFunction::sum($DB::quoteName('TABLE.capacity')) . ' / ' .
+                QueryFunction::count($DB::quoteName('TABLE.id')) . ') * ' .
+                QueryFunction::count(expression: $DB::quoteName('TABLE.id'), distinct: true),
             'nometa'             => true, // cannot GROUP_CONCAT a SUM
         ];